requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
sqlalchemy[asyncio]==2.0.25
boto3==1.34.14
openai==1.10.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0
python-dotenv==1.0.0
aiohttp==3.9.3
aiofiles==23.2.1
